% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggInvFn}
\alias{aggInvFn}
\title{The function calculate targeting score of miRNA w.r.t to a cluster
of pathways via inverse normal method}
\usage{
aggInvFn(enriches, pathways, isSelector = TRUE, thresh = NULL)
}
\arguments{
\item{enriches}{a table of miRNA pathway enrichments. Universe}

\item{pathways}{queried pathways. e.g. cluster pathways}

\item{isSelector}{internal argument}

\item{thresh}{internal argument}
}
\value{
a  scoring of miRNAs in a cluster of pathways
}
\description{
The function calculate targeting score of miRNA w.r.t to a cluster
of pathways via inverse normal method
}
\keyword{internal}
