% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alignToUniverse}
\alias{alignToUniverse}
\title{function to align a list of sets and a reference universe}
\usage{
alignToUniverse(pathwaySets, universe)
}
\arguments{
\item{pathwaySets}{a list of sets}

\item{universe}{all set  elements must be a subset of universe}
}
\value{
a list of sets, aligned to universe
}
\description{
function to align a list of sets and a reference universe
}
