% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COV_TCGA_LIHC}
\alias{COV_TCGA_LIHC}
\title{Covariates/metadata of The Cancer Genome Atlas Liver Hepatocellular 
Carcinoma gene expression data.}
\format{
A data frame with 417 rows and 73 variables.
}
\source{
\url{https://doi.org/10.1016/j.cell.2017.05.046}
}
\usage{
COV_TCGA_LIHC
}
\description{
This dataset is filtered to contain only sample with miRNA and mRNA gene
expression profiles.
}
\keyword{datasets}
