% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-DifferentialPathwayAnalysis.R
\name{DifferentialPathwayAnalysis}
\alias{DifferentialPathwayAnalysis}
\title{Differential Expression Analysis For Pathways}
\usage{
DifferentialPathwayAnalysis(
  genes.counts,
  pathways,
  covariates,
  condition,
  adjust.covars = NULL,
  covariates.correction = FALSE,
  quantile.norm = FALSE,
  out.dir = "",
  save.RDS.name = NULL,
  id = "ENSEMBL",
  de.genes = NULL,
  min.path.size = 10,
  method = "x2",
  trim = 0.025,
  genes.counts.log = TRUE,
  contrast.conds = NA
)
}
\arguments{
\item{genes.counts}{Gene counts, rows refer to genes and columns to samples.}

\item{pathways}{Pathways table, containing pathway names and genes with id
specified.}

\item{covariates}{Covariates/metadata file; rows matches the columns of
gene.counts.}

\item{condition}{Condition to be examined (tumor vs normal etc); must exist
in covariates column.}

\item{adjust.covars}{Adjustment covariates like batch; if NULL,
no adjustments performed.}

\item{covariates.correction}{If TRUE, performs covariates detection and
correction; requires **adjust.covars**; (limma).}

\item{quantile.norm}{If TRUE, performs quantile normalization on pathway
summary statistics; from *preprocess* package.}

\item{out.dir}{Output directory.}

\item{save.RDS.name}{If not NULL, saves output as RDS using save name,
if NULL, does not save output.}

\item{id}{ID matching genes to pathways; rownames of gene.counts.}

\item{de.genes}{If not NULL, add t-scores to pathways summary statistics;
filter by genes t-scores.}

\item{min.path.size}{Minimum pathway size.}

\item{method}{Define method to use for pathway summary statistics;
specifications in documentations.}

\item{trim}{Filter pathways with mean less than trim threshold
in pathway summary statistics.}

\item{genes.counts.log}{If TRUE, log(genes.counts).}

\item{contrast.conds}{Provide a contrast expression to be used in Limma
comparison. This is necessary if you have more than two levels in the
condition covariate.}
}
\value{
List containing differentially expressed pathways as DEP and pathway
  summary statistics as pathwaySummaryStats.
}
\description{
Performs differential expression analysis for pathways using LIMMA package
with gene counts
}
