% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enrichAllPairs}
\alias{enrichAllPairs}
\title{Pairwise enrichment analysis between two given lists of sets}
\usage{
enrichAllPairs(mirSets, pathwaySets, pathsRef, numCores)
}
\arguments{
\item{mirSets}{a list of targets of miRNAs}

\item{pathwaySets}{a list of pathways}

\item{pathsRef}{universe of genes.}

\item{numCores}{number of cores to calculate the results.}
}
\value{
enrichment analysis results
}
\description{
Pairwise enrichment analysis between two given lists of sets
}
