% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Obtain Design Matrix}
\usage{
getDesignMatrix(covariatesDataFrame, Intercept = TRUE, RELEVELS = list())
}
\arguments{
\item{covariatesDataFrame}{Dataframe of covariates.}

\item{Intercept}{Intercept in the linear model.}

\item{RELEVELS}{TBA.}
}
\value{
List containing a design matrix.
}
\description{
Modified from covariates pipeline of Menachem Former. Imported from
\url{https://github.com/th1vairam/CovariateAnalysis}
}
