% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDiffExpTable}
\alias{getDiffExpTable}
\title{function to get a DE table}
\usage{
getDiffExpTable(expMat, designMat, contrastsName)
}
\arguments{
\item{expMat}{an expression matrix}

\item{designMat}{a design Matrix}

\item{contrastsName}{the contrast to perform}
}
\value{
a table of differential expression
}
\description{
function to get a DE table
}
