% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getResidual}
\alias{getResidual}
\title{function to get residuals with respect to a set of covariates}
\usage{
getResidual(covariates, adjustCovars, pathwaySummaryStats)
}
\arguments{
\item{covariates}{a covariate dataframe.}

\item{adjustCovars}{covariates to adjust for}

\item{pathwaySummaryStats}{an expression matrix}
}
\value{
a matrix of adjusted expression
}
\description{
function to get residuals with respect to a set of covariates
}
