% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gscExample}
\alias{gscExample}
\title{Example genesets from MSigDB}
\format{
A GeneSet Collection object containing two genesets.
}
\source{
\url{http://www.gsea-msigdb.org/gsea/index.jsp}
}
\usage{
data(gscExample)
}
\description{
Example genesets from MSigDB
}
\examples{
data(gscExample)
}
\keyword{datasets}
