% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-functions.R
\name{lancaster.fn}
\alias{lancaster.fn}
\title{The function calculate targeting score of miRNA w.r.t to a cluster 
of pathways via lancaster aggregation method.}
\usage{
lancaster.fn(enriches, pathways, is.selector, thresh = NULL)
}
\arguments{
\item{enriches}{a table of miRNA pathway enrichments. Universe}

\item{pathways}{queried pathways. e.g. cluster pathways}

\item{is.selector}{internal argument}

\item{thresh}{internal argument}
}
\value{
a  scoring of miRNAs in a cluster of pathways
}
\description{
The function calculate targeting score of miRNA w.r.t to a cluster 
of pathways via lancaster aggregation method.
}
