% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-functions.R
\name{linColumnFinder}
\alias{linColumnFinder}
\title{Function imported from https://github.com/th1vairam/CovariateAnalysis
Modified from http://stackoverflow.com/questions/13088770/how-to-write-linearly-dependent-column-in-a-matrix-in-terms-of-linearly-independ
Function to find linearly dependednt columns of a matrix}
\usage{
linColumnFinder(mat)
}
\arguments{
\item{mat}{an input design matrix.}
}
\value{
a list of independent columns
}
\description{
Function imported from https://github.com/th1vairam/CovariateAnalysis
Modified from http://stackoverflow.com/questions/13088770/how-to-write-linearly-dependent-column-in-a-matrix-in-terms-of-linearly-independ
Function to find linearly dependednt columns of a matrix
}
