% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniEnrich}
\alias{miniEnrich}
\title{A table of miRNA-pathway enrichment.
This is a reduced representation of TargetScan
For full table use miRNAPathwayEnrichment function in the package
along with msigdb_c2 and targetScan_03 datasets.}
\format{
A dataframe of 13170 rows and 10 variables:
\describe{
  \item{x}{A targetscan annotated miRNA}
  \item{y}{An MSigDB annotated pathway}
  \item{pval}{Fisher's Exact test p-value}
}
}
\usage{
data(miniEnrich)
}
\description{
A table of miRNA-pathway enrichment.
This is a reduced representation of TargetScan
For full table use miRNAPathwayEnrichment function in the package
along with msigdb_c2 and targetScan_03 datasets.
}
\examples{
data(miniEnrich)
}
\keyword{datasets}
