% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_c2}
\alias{msigdb_c2}
\title{Canonical pathways from Molecular Signatures Database, MsigDb V6.2}
\format{
A list of 1329 pathways
}
\source{
\url{http://www.gsea-msigdb.org/gsea/index.jsp}
}
\usage{
data(msigdb_c2)
}
\description{
Canonical pathways from Molecular Signatures Database, MsigDb V6.2
}
\examples{
data(msigdb_c2)
}
\keyword{datasets}
