% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_gene_table}
\alias{path_gene_table}
\title{A table of gene-pathway association. based on the pathways of MSigDB.}
\format{
A matrix with 3 columns and 76926 rows:
\describe{
  \item{Pathway}{An MSigDB annotated pathway}
  \item{ENTREZID}{The ENTREZID of a gene belonging to the pathway}
  \item{ENSEMBL}{The ENSEMBL of a gene belonging to the pathway}
}
}
\usage{
data(path_gene_table)
}
\description{
A table of gene-pathway association. based on the pathways of MSigDB.
}
\keyword{datasets}
