% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Path_Summary}
\alias{Path_Summary}
\title{The function to generate pathway summary statistics}
\usage{
Path_Summary(
  exprs.mat,
  pathway.ref,
  id,
  z.normalize = F,
  method = F,
  de.genes = NULL,
  trim = 0,
  t.scores = NULL
)
}
\arguments{
\item{exprs.mat}{a gene expression matrix with rownames as genes and samples
as columns.}

\item{pathway.ref}{a table of pathway-gene associations. Created from 
\code{\link{Pathway_Gene_Tab}} function}

\item{id}{the gene annotation type in the rowname of gene expression data}

\item{z.normalize}{normalization of pathway summary score}

\item{method}{choice of how to sumamrize gene ranks into pathway statistics.}

\item{de.genes}{a list of differentially expressed genes along with
t-scores. only necessary if working on Top 50% summary method.}

\item{trim}{percentage of top and bottom ranked genes to be excluded from
pathway summary statistics}

\item{t.scores}{argument for-top-50-percent-genes method}
}
\value{
PathExp a table of pathway activity profiles per sample.
}
\description{
The function to generate pathway summary statistics
}
