% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Pathway_Gene_Tab}
\alias{Pathway_Gene_Tab}
\title{Generates a table of pathways and genes associations}
\usage{
Pathway_Gene_Tab(path.address = F, out.dir = F)
}
\arguments{
\item{path.address}{address to an RDS file cointaning list of pathways
each element is a list of genes similar to GMT format.}

\item{out.dir}{address to save an RDS for a table of pathway-gene association}
}
\value{
pathExpTab a table of pathway-gene association
}
\description{
Generates a table of pathways and genes associations
}
