% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwayClustersLIHC}
\alias{pathwayClustersLIHC}
\title{An output of the mappingPathwayClusters method. containing mapped
differentially activated pathways from TCGA LIHC dataset to PCxN network
using Louvain clustering algorithm}
\format{
A list of lists containing a list of cluster assignment to 
    pathways. An igraph object of the mapped PCxN.
}
\source{
\url{http://pcxn.org:8080/}
}
\usage{
data(pathwayClustersLIHC)
}
\description{
An output of the mappingPathwayClusters method. containing mapped
differentially activated pathways from TCGA LIHC dataset to PCxN network
using Louvain clustering algorithm
}
\examples{
data(pathwayClustersLIHC)
}
\keyword{datasets}
