% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathwayGeneTab}
\alias{pathwayGeneTab}
\title{Pathway-Gene Associations}
\usage{
pathwayGeneTab(pathAdress = NA, pathwayList = NA, outDir = NA)
}
\arguments{
\item{pathAdress}{Address to an RDS file containing list of pathways where
each element is a list of genes similar to GMT format.}

\item{pathwayList}{If you wish to use a list of pathways instead of a file
use this argument instead. The list must contain no NA values.}

\item{outDir}{Address to save an RDS for a table of pathway-gene association}
}
\value{
pathExpTab Table of pathway-gene association.
}
\description{
Generates a table of pathways and genes associations.
}
