% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathwaySummary}
\alias{pathwaySummary}
\title{Pathway Summary Statistics}
\usage{
pathwaySummary(
  exprsMat,
  pathwayRef,
  id = "ENSEMBL",
  zNormalize = FALSE,
  method = FALSE,
  deGenes = NULL,
  trim = 0,
  tScores = NULL
)
}
\arguments{
\item{exprsMat}{Gene expression matrix with row names as genes and samples
as columns.}

\item{pathwayRef}{Table of pathway-gene associations. Created from
\code{\link{pathwayGeneTab}} function.}

\item{id}{Gene annotation type in the row name of gene expression data.}

\item{zNormalize}{Normalization of pathway summary score.}

\item{method}{Choice of how to summarize gene ranks into pathway statistics.}

\item{deGenes}{List of differentially expressed genes along with t-scores.
Only necessary if working on Top 50\% summary method.}

\item{trim}{Percentage of top and bottom ranked genes to be excluded from
pathway summary statistics.}

\item{tScores}{Argument for-top-50-percent-genes method.}
}
\value{
pathExp Table of pathway activity profiles per sample.
}
\description{
Generates a table of pathway activity profiles per sample
}
