% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-functions.R
\name{pCut.fn}
\alias{pCut.fn}
\title{The function to count the number of enriched pathways for each miRNA}
\usage{
pCut.fn(enriches, pathways, is.selector, thresh = 0.05)
}
\arguments{
\item{enriches}{a table of miRNA pathway enrichments. Universe}

\item{pathways}{queried pathways. e.g. cluster pathways}

\item{is.selector}{internal argument}

\item{thresh}{threshold from p-value cut-off}
}
\value{
a p-value based scoring of miRNAs in a cluster of pathways
}
\description{
The function to count the number of enriched pathways for each miRNA
}
