% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pCutFn}
\alias{pCutFn}
\title{Score miRNAs In a Cluster Of Pathways}
\usage{
pCutFn(enriches, pathways, isSelector, thresh = 0.05)
}
\arguments{
\item{enriches}{Table of miRNA pathway enrichments.}

\item{pathways}{Queried pathways, e.g. cluster pathways.}

\item{isSelector}{Internal argument.}

\item{thresh}{Threshold from p-value cut-off.}
}
\value{
P-value based scoring of miRNAs in a cluster of pathways.
}
\description{
The function to count the number of enriched pathways for each miRNA.
}
