% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcxn}
\alias{pcxn}
\title{The pathway correlation network of canonical pathways from the MSigDB.}
\format{
A matrix with 6 columns and 882456 rows:
\describe{
  \item{Pathway.A}{An MSigDB annotated pathway}
  \item{Pathway.B}{The ENTREZID of a gene belonging to the pathway}
  \item{Overlap.coefficient}{Jaccard index of the two pathways}
  \item{PathCor}{Correlation coefficient of the two pathways}
  \item{p.value}{correlation p.value}
  \item{p.Adjust}{FDR corrected p.value of correlation}
}
}
\source{
\url{https://pcxn.org}
}
\usage{
pcxn
}
\description{
Each row represents two pathways, their overlap, their correlation
coefficient, and respective p-value of correlation.
}
\details{
Please cite the appropriate publication if you use pcxn.

https://doi.org/10.1371/journal.pcbi.1006042
}
\keyword{datasets}
