% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcxnToNet}
\alias{pcxnToNet}
\title{Creates a network out of pcxn table}
\usage{
pcxnToNet(pcxn, edgeFDR, correlationCutOff, weighted)
}
\arguments{
\item{pcxn}{pathways network edge list of pathways}

\item{edgeFDR}{FDR threshold for pathway-pathway adjusted p-values;
filter edges with adjusted p-values less than given threshold}

\item{correlationCutOff}{cut-off threshold for pathway-pathway correlation;
filter pathways with correlation less than given threshold}

\item{weighted}{True if you wish to include correlation weights in clustering}
}
\value{
enrichment analysis results
}
\description{
Creates a network out of pcxn table
}
