% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-miRNAPrioritization.R
\name{PrioritizeMicroRNA}
\alias{PrioritizeMicroRNA}
\title{Outputs a table of miRNA (ordered) with respective p-values derived from method for prioritization}
\usage{
PrioritizeMicroRNA(
  enriches0,
  pathway.clusters,
  method = "AggInv",
  method.thresh = NULL,
  mir.path.fdr.thresh = 0.25,
  top.clust = 2,
  samp.rate = 1000,
  out.dir = "",
  data.dir = "",
  save.sampling = T,
  run.jack.knife = T,
  save.jack.knife = F,
  num.cores = 1,
  save.csv = T,
  prefix = ""
)
}
\arguments{
\item{pathway.clusters}{pathway clusters, obtained from MappingPathwaysClusters}

\item{method}{a vector of methods (pCut, AggInv, AggLog, sumz, sumlog)}

\item{method.thresh}{a vector of methods threshold for each method in method, if NULL use default thresh values in method}

\item{top.clust}{top x clusters to perform miRNA prioritization on}

\item{samp.rate}{sampling rate for CLT}

\item{out.dir}{output directory}

\item{data.dir}{data directory}

\item{save.sampling}{if T, saves sampling data as RDS for each cluster in top.clust in data.dir}

\item{save.jack.knife}{if T, saves jack-knifed sampling data as RDS for each cluster in top.clust in data.dir}

\item{num.cores}{number of cores available}

\item{save.csv}{if T, saves csv file for each cluster in top.clust in out.dir}

\item{prefix}{prefix for all saved data}

\item{enriches}{miRNA-pathway enrichment dataset, obtained from miRNAPathwayEnrichment}

\item{mir.path.fdr.selection}{FDR threshold for mir-pathway adjusted p-values ; filter edges with adjusted p-values less than given threshold}
}
\value{
table of miRNA and p-values, each row contains a mirna and its associated p-values from the methods
}
\description{
Outputs a table of miRNA (ordered) with respective p-values derived from method for prioritization
}
