% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reportEnrichment}
\alias{reportEnrichment}
\title{Publication-ready miRNA-Pathway Enrichment table}
\usage{
reportEnrichment(enrichmentTable)
}
\arguments{
\item{enrichmentTable}{Outputs from [miRNAPathwayEnrichment()] function}
}
\value{
A summarized miRNA-Pathway enrichment table
}
\description{
This function summarizes the outputs
}
\examples{
data(msigdb_c2)
data(targetScan_03)
eTab <- miRNAPathwayEnrichment(targetScan_03[1:20],msigdb_c2[1:20])

repTab <- reportEnrichment(eTab)
}
