% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samplingDataBase}
\alias{samplingDataBase}
\title{Outputs a table of sampling data(rows are miRNA and cols are samples)}
\usage{
samplingDataBase(
  enrichNull,
  selector,
  samp.rate,
  fn,
  n_paths,
  samplingDataFile,
  jackKnife = FALSE,
  saveSampling,
  numCores = 1
)
}
\arguments{
\item{enrichNull}{Enrichment dataset with x (miRNA), y (pathway) and pval
(probability of observing x in pathway cluster).}

\item{selector}{Table with x(miRNA) in pathway cluster.}

\item{samp.rate}{Sampling rate.}

\item{fn}{Methodology function.}

\item{n_paths}{Number of pathways in pathway cluster.}

\item{samplingDataFile}{If file exists, load. Else, perform random sampling}

\item{jackKnife}{If TRUE, conduct sampling with one less pathway, used for
jack knifing}

\item{saveSampling}{If TRUE, data is saved.}

\item{numCores}{number of cores used}
}
\value{
Outputs of sampling data.
}
\description{
Outputs a table of sampling data(rows are miRNA and cols are samples)
}
