% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumlogCoverFn}
\alias{sumlogCoverFn}
\title{Internal function for modification of prioritization.}
\usage{
sumlogCoverFn(selector, coverName)
}
\arguments{
\item{selector}{a prioritzation table}

\item{coverName}{a new column name}
}
\value{
an updated scoring of miRNAs in a cluster of pathways
}
\description{
Internal function for modification of prioritization.
}
