% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumz.cover.fn}
\alias{sumz.cover.fn}
\title{Internal function for modification of prioritization.}
\usage{
sumz.cover.fn(selector, cover.name)
}
\arguments{
\item{selector}{a prioritzation table}

\item{cover.name}{a new column name}
}
\value{
an updated scoring of miRNAs in a cluster of pathways
}
\description{
Internal function for modification of prioritization.
}
