% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumzFn}
\alias{sumzFn}
\title{The function calculate targeting score of miRNA w.r.t to a cluster
of pathways via sumz aggregation method.}
\usage{
sumzFn(enriches, pathways, isSelector, thresh = NULL)
}
\arguments{
\item{enriches}{a table of miRNA pathway enrichments. Universe}

\item{pathways}{queried pathways. e.g. cluster pathways}

\item{isSelector}{internal argument}

\item{thresh}{internal argument}
}
\value{
a  scoring of miRNAs in a cluster of pathways
}
\description{
The function calculate targeting score of miRNA w.r.t to a cluster
of pathways via sumz aggregation method.
}
\keyword{internal}
