% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tableFromGSC}
\alias{tableFromGSC}
\title{Pathway-Gene Associations from GeneSet collections}
\usage{
tableFromGSC(gsCollection, fromType = "ENTREZID", toType = "ENSEMBL")
}
\arguments{
\item{gsCollection}{An GSEABase gene set collection object}

\item{fromType}{gene annotation type used in your input data}

\item{toType}{gene annotation type to be produced in the output}
}
\value{
A table of pathway-gene associations
}
\description{
This function enables to utilize MSigDB packages and GSEABase objects
  to incorporate customized genesets into PanomiR.
}
\examples{
data(gscExample)
tableFromGSC(gscExample)
}
