% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetScan_03}
\alias{targetScan_03}
\title{A processed list of miRNA target gene sets from the TargetScan dataset. 
Each list item is a list of genes targeted by the respective miRNA family}
\format{
A list of 439 items
}
\source{
\url{http://www.targetscan.org/vert_72/}
}
\usage{
targetScan_03
}
\description{
The interactions are filtered to only human interactions.
}
\details{
The interactions are filtered to have a 
Cumulative weighted context++ score of < -0.3
}
\keyword{datasets}
