% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_LIHC}
\alias{TCGA_LIHC}
\title{The gene expression data The Cancer Genome Atlas Liver Hepatocellular
Carcinoma}
\format{
A matrix with 14120 rows and 417 columns.
}
\source{
\url{https://doi.org/10.1016/j.cell.2017.05.046}
}
\usage{
TCGA_LIHC
}
\description{
This dataset is filtered to contain only sample with miRNA and mRNA gene
expression profiles.
}
\keyword{datasets}
