% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Function imported from https://github.com/th1vairam/CovariateAnalysis
Function to obtain desing matrix (modified from covairates pipeline of Menachem Former)}
\usage{
getDesignMatrix(covariatesDataFrame, Intercept = T, RELEVELS = list())
}
\arguments{
\item{covariatesDataFrame}{a datadframe of covariates.}

\item{Intercept}{intercept in the linear model}

\item{RELEVELS}{TBA}

\item{pathways}{Pathways in pathway cluster.}
}
\value{
a list containing a design matrix
}
\description{
Function imported from https://github.com/th1vairam/CovariateAnalysis
Function to obtain desing matrix (modified from covairates pipeline of Menachem Former)
}
