% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Obtain Design Matrix}
\usage{
getDesignMatrix(covariatesDataFrame, intercept = TRUE, reLevels = list())
}
\arguments{
\item{covariatesDataFrame}{Dataframe of covariates.}

\item{intercept}{intercept in the linear model.}

\item{reLevels}{TBA.}
}
\value{
List containing a design matrix.
}
\description{
Modified from covariates pipeline of Menachem Former. Imported from
\url{https://github.com/th1vairam/CovariateAnalysis}
}
\examples{
data(iris)
getDesignMatrix(iris)
}
