% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linColumnFinder}
\alias{linColumnFinder}
\title{Function imported from https://github.com/th1vairam/CovariateAnalysis
Modified from http://stackoverflow.com/questions/13088770/
Function to find linearly dependednt columns of a matrix}
\usage{
linColumnFinder(mat)
}
\arguments{
\item{mat}{an input design matrix.}
}
\value{
a list of independent columns
}
\description{
Function imported from https://github.com/th1vairam/CovariateAnalysis
Modified from http://stackoverflow.com/questions/13088770/
Function to find linearly dependednt columns of a matrix
}
\examples{
data("iris")
designMat <- getDesignMatrix(iris)
linColumnFinder(designMat$design)
}
