% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-MappingPathwaysClusters.R
\name{MappingPathwaysClusters}
\alias{MappingPathwaysClusters}
\title{Outputs a table with pathways and their respective clusters}
\usage{
MappingPathwaysClusters(
  pcxn,
  de.paths,
  clust.fn = NULL,
  edge.fdr.thresh = 0.05,
  cor.thresh = 0.316,
  path.fdr.thresh = 0.05,
  top.paths = 200,
  seed = 2,
  plot = T,
  subplot = T,
  top.clusts = 2,
  prefix = "",
  out.dir = "",
  save.csv.name = NULL,
  weighted = F
)
}
\arguments{
\item{pcxn}{pathways network (edge list of pathways)}

\item{de.paths}{differential expressed pathways, obtained from
*DifferentialPathwayAnalysis*}

\item{clust.fn}{clustering algorithm}

\item{edge.fdr.thresh}{FDR threshold for pathway-pathway adjusted p-values;
filter edges with adjusted p-values less than given threshold}

\item{cor.thresh}{correlation threshold for pathway-pathway correlation;
filter pathways with correlation less than given threshold}

\item{path.fdr.thresh}{FDR threshold for DE pathways adjusted p-values;
filter pathways with adjusted p-values less than given threshold}

\item{top.paths}{use only top x paths; if NULL, use all paths}

\item{seed}{set seed}

\item{plot}{if T, store graph plot in Figures directory of plots}

\item{subplot}{if T, store inidividual clusters plots and connected plots
in Figures directory of plots}

\item{top.clusts}{plot figures for top x clusters}

\item{prefix}{add prefix to plots}

\item{out.dir}{output directory}

\item{save.csv.name}{if not NULL, saves output as csv using save name}

\item{weighted}{True if you wish to include correlation weights in clustering}
}
\value{
a table with each row containing a pathway and its respective cluster
}
\description{
Outputs a table with pathways and their respective clusters
}
