% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-functions.R
\name{methodProbBase}
\alias{methodProbBase}
\title{Outputs a table with col x (miRNA), probability of
observing k (depending on methodology)
against a random distribution and cover of methodology}
\usage{
methodProbBase(sampling.data, selector, m, n_paths = 100, cover.fn = NULL)
}
\arguments{
\item{sampling.data}{Random distribution data}

\item{selector}{Table with x(miRNA) in pathway cluster and observed
k (depending on methodology).}

\item{m}{method name.}

\item{cover.fn}{Cover of methodology function.}
}
\value{
Outputs a new selector table with col x, pval and cover.
}
\description{
Outputs a table with col x (miRNA), probability of
observing k (depending on methodology)
against a random distribution and cover of methodology
}
