% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-miRNAPathwayEnrichment.R
\name{miRNAPathwayEnrichment}
\alias{miRNAPathwayEnrichment}
\title{Outputs enrichment probability of miRNAs based on pathway clusters}
\usage{
miRNAPathwayEnrichment(
  mir.sets,
  pathways.sets,
  genes.selection = NULL,
  mir.selection = NULL,
  from.id = "ENSEMBL",
  to.id = "ENTREZID",
  min.path.size = 9,
  num.cores = 1,
  out.dir = "",
  save.RDS.name = NULL
)
}
\arguments{
\item{mir.sets}{a table of miRNAs and a list of their interactions with genes in ENTREZ ID}

\item{pathways.sets}{a table of pathways and a list of their interactions with genes in ENTREZ ID}

\item{mir.selection}{a table of miRNA names; if not NULL, select only miRNAs from given table}

\item{from.id}{id of genes in genes.selection}

\item{to.id}{id of genes used in pcxn and pathways set}

\item{min.path.size}{filter out pathways with sets less than given value}

\item{num.cores}{number of cores available}

\item{out.dir}{output directory}

\item{save.RDS.name}{if not NULL, saves output as RDS using save name}

\item{gene.selection}{a table of genes with dtype; if not NULL, select only genes from given table}
}
\value{
table of enrichment, each row contains mirna-pathway and its enrichment p-values
}
\description{
Outputs enrichment probability of miRNAs based on pathway clusters
}
