% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-miRNAPathwayEnrichment.R
\name{miRNAPathwayEnrichment}
\alias{miRNAPathwayEnrichment}
\title{Enrichment Probability Of miRNAs}
\usage{
miRNAPathwayEnrichment(
  mir.sets,
  pathways.sets,
  genes.selection = NULL,
  mir.selection = NULL,
  from.id = "ENSEMBL",
  to.id = "ENTREZID",
  min.path.size = 9,
  num.cores = 1,
  out.dir = "",
  save.RDS.name = NULL
)
}
\arguments{
\item{mir.sets}{Table of miRNAs and a list of their interactions with
genes in ENTREZ ID.}

\item{pathways.sets}{Table of pathways and a list of their interactions
with genes in ENTREZ ID.}

\item{mir.selection}{Table of miRNA names; if not NULL, select only miRNAs
from given table.}

\item{from.id}{ID of genes in genes.selection.}

\item{to.id}{ID of genes used in pcxn and pathways set.}

\item{min.path.size}{Filter out pathways with sets less than given value.}

\item{num.cores}{Number of CPU cores to use, must be at least one.}

\item{out.dir}{Output directory.}

\item{save.RDS.name}{If not NULL, saves output as RDS using save name.}

\item{gene.selection}{Table of genes with dtype; if not NULL, select only
genes from a given table.}
}
\value{
Table of enrichment, each row contains mirna-pathway and its 
  enrichment p-values
}
\description{
Outputs enrichment probability of miRNAs based on pathway clusters.
}
