% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Path_Summary}
\alias{Path_Summary}
\title{Pathway Summary Statistics}
\usage{
Path_Summary(
  exprs.mat,
  pathway.ref,
  id = "ENSEMBL",
  z.normalize = FALSE,
  method = FALSE,
  de.genes = NULL,
  trim = 0,
  t.scores = NULL
)
}
\arguments{
\item{exprs.mat}{Gene expression matrix with row names as genes and samples
as columns.}

\item{pathway.ref}{Table of pathway-gene associations. Created from
\code{\link{Pathway_Gene_Tab}} function.}

\item{id}{Gene annotation type in the row name of gene expression data.}

\item{z.normalize}{Normalization of pathway summary score.}

\item{method}{Choice of how to summarize gene ranks into pathway statistics.}

\item{de.genes}{List of differentially expressed genes along with t-scores.
Only necessary if working on Top 50\% summary method.}

\item{trim}{Percentage of top and bottom ranked genes to be excluded from
pathway summary statistics.}

\item{t.scores}{Argument for-top-50-percent-genes method.}
}
\value{
PathExp Table of pathway activity profiles per sample.
}
\description{
Generates a table of pathway activity profiles per sample
}
