% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Pathway_Gene_Tab}
\alias{Pathway_Gene_Tab}
\title{Pathway-Gene Associations}
\usage{
Pathway_Gene_Tab(path.address = NA, pathway.list = NA, out.dir = NA)
}
\arguments{
\item{path.address}{Address to an RDS file containing list of pathways where
each element is a list of genes similar to GMT format.}

\item{out.dir}{Address to save an RDS for a table of pathway-gene association}
}
\value{
pathExpTab Table of pathway-gene association.
}
\description{
Generates a table of pathways and genes associations.
}
