% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-miRNAPrioritization.R
\name{PrioritizeMicroRNA}
\alias{PrioritizeMicroRNA}
\title{Prioritize miRNA}
\usage{
PrioritizeMicroRNA(
  enriches0,
  pathway.clusters,
  method = "AggInv",
  method.thresh = NULL,
  mir.path.fdr.thresh = 0.25,
  top.clust = 2,
  samp.rate = 1000,
  out.dir = "",
  data.dir = "",
  save.sampling = TRUE,
  run.jack.knife = TRUE,
  save.jack.knife = FALSE,
  num.cores = 1,
  save.csv = TRUE,
  prefix = ""
)
}
\arguments{
\item{enriches0}{miRNA-pathway enrichment dataset obtained from
miRNAPathwayEnrichment.}

\item{pathway.clusters}{Pathway clusters, obtained from
MappingPathwaysClusters.}

\item{method}{Vector of methods (pCut, AggInv, AggLog, sumz, sumlog).}

\item{method.thresh}{Vector of methods threshold for each method in method,
if NULL use default thresh values in method.}

\item{mir.path.fdr.thresh}{for calculating miRNA-pathway hits in the input
cluster based on standard enrichment analysis.}

\item{top.clust}{Top x clusters to perform miRNA prioritization on.}

\item{samp.rate}{Sampling rate for CLT.}

\item{out.dir}{Output directory.}

\item{data.dir}{Data directory.}

\item{save.sampling}{If TRUE, saves sampling data as RDS for each cluster in
top.clust in data.dir.}

\item{run.jack.knife}{If TRUE, jacknifing will be performed.}

\item{save.jack.knife}{If TRUE, saves jack-knifed sampling data as RDS for each
cluster in top.clust in data.dir.}

\item{num.cores}{Number of CPU cores to use, must be at least one.}

\item{save.csv}{If TRUE, saves CSV file for each cluster in top.clust in
out.dir.}

\item{prefix}{Prefix for all saved data.}
}
\value{
Table of miRNA and p-values, each row contains a miRNA and its
  associated p-values from the methods.
}
\description{
Outputs a table of miRNA (ordered) with respective p-values derived from
  method for prioritization
}
