% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samplingDataBase}
\alias{samplingDataBase}
\title{Outputs a table of sampling data(rows are miRNA and cols are samples)}
\usage{
samplingDataBase(
  enrich.null,
  selector,
  samp.rate,
  fn,
  n_paths,
  sampling.data.file,
  jack.knife = FALSE,
  save.sampling,
  num.cores = 1
)
}
\arguments{
\item{enrich.null}{Enrichment dataset with x (miRNA), y (pathway) and pval
(probability of observing x in pathway cluster).}

\item{selector}{Table with x(miRNA) in pathway cluster.}

\item{samp.rate}{Sampling rate.}

\item{fn}{Methodology function.}

\item{n_paths}{Number of pathways in pathway cluster.}

\item{sampling.data.file}{If file exists, load file. Else, perform random sampling}

\item{jack.knife}{If TRUE, conduct sampling with one less pathway, used for
jack knifing}

\item{save.sampling}{If TRUE, data is saved.}

\item{num.cores}{number of cores used}
}
\value{
Outputs of sampling data.
}
\description{
Outputs a table of sampling data(rows are miRNA and cols are samples)
}
